+++ frontend/src/pages/user/Password/usePassword.ts
@@ -1,0 +1,47 @@
+import { useState } from 'react';
+import { useNavigate, useParams } from 'react-router-dom';
+
+import apis from '@/apis';
+
+const usePassword = () => {
+  const navigate = useNavigate();
+
+  const [isActiveSubmit, setIsActiveSubmit] = useState(false);
+  const { hostId } = useParams();
+
+  const setToken = async (password: string) => {
+    const { token } = await apis.postPassword({ hostId, password });
+    localStorage.setItem('user', token);
+  };
+
+  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    if (!isActiveSubmit) return;
+
+    const form = e.target as HTMLFormElement;
+    const { value: password } = form[0] as HTMLInputElement;
+
+    try {
+      await setToken(password).then(() => {
+        navigate(`/enter/${hostId}/spaces`);
+      });
+    } catch (err) {
+      alert('비밀번호를 확인해주세요.');
+    }
+  };
+
+  const handelChangeInput = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { value } = e.target as HTMLInputElement;
+
+    const isExistValue = value.length > 0;
+    setIsActiveSubmit(isExistValue);
+  };
+
+  return {
+    isActiveSubmit,
+    onSubmit: handleSubmit,
+    onChange: handelChangeInput,
+  };
+};
+
+export default usePassword;
